refactor(xatu): make Xatu integration optional via build tag embedded
Moves Xatu code behind the `embedded` build tag so the dependency is
only compiled when the tag is present. This keeps the default build
lean while still allowing full Xatu functionality when needed.

+++ node/xatu/datasource.go
@@ -14,6 +14,8 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
+//go:build embedded
+
 package xatu
 
 import (
@@ -249,9 +251,9 @@ func (s *Service) DebugTraceTransaction(
 }
 
 // ChainID returns the chain ID.
-func (s *Service) ChainID() int32 {
+func (s *Service) ChainID() int64 {
 	if s.chainConfig.ChainID != nil {
-		return int32(s.chainConfig.ChainID.Int64())
+		return s.chainConfig.ChainID.Int64()
 	}
 
 	return 1
